refactor: product.js 분리
